Created a new class for the AWS EFS file system deployment feature. This will allow us to easily choose which ConfigBuilder to use. The tag "aws-efs" must be passed to CIV in order to use this new class.
